fix all warnings and compilation errors with -Wall -Wextra -Wmisleading-indentation -Wunused -Wuninitialized -Wshadow -Wconversion on all platforms
directly import posix filesystem apis with extern function since some operating systems do not correctly provide renameat or other filesystem syscalls

+++ include/fast_io_core_impl/integers/sto/sto_contiguous.h
@@ -254,7 +254,7 @@ inline std::uint32_t detect_length(char unsigned const* buffer) noexcept
 	__m128i const mask = _mm_cmplt_epi8(t0, _mm_set1_epi8(-118));
 	std::uint16_t v{static_cast<std::uint16_t>(_mm_movemask_epi8(mask))};
 #endif
-	return std::countr_one(v);
+	return static_cast<std::uint32_t>(std::countr_one(v));
 }
 
 template<bool char_execharset>
